Move compass integration jobs to separate file
