+++ prow/jobs/kyma/kyma-integration.yaml
@@ -14,11 +14,6 @@ kyma_ref: &kyma_ref
   repo: kyma
   path_alias: github.com/kyma-project/kyma
 
-compass_ref: &compass_ref
-  org: kyma-incubator
-  repo: compass
-  path_alias: github.com/kyma-incubator/compass
-
 vm_job_template: &vm_job_template
   skip_report: false # from documentation: SkipReport skips commenting and setting status on GitHub.
   decorate: true
@@ -532,48 +527,6 @@ presubmits: # runs on PRs
     extra_refs:
     - <<: *test_infra_ref
       base_ref: master
-
-  - name: pre-master-kyma-gke-compass-integration-on-demand
-    optional: true
-    decorate: true
-    path_alias: github.com/kyma-project/kyma
-    skip_report: false
-    max_concurrency: 10
-    spec:
-      containers:
-        - image: eu.gcr.io/kyma-project/test-infra/kyma-cluster-infra:v20190528-8897828
-          securityContext:
-            privileged: true
-          command:
-            - "bash"
-          args:
-            - "-c"
-            - "${KYMA_PROJECT_DIR}/test-infra/prow/scripts/cluster-integration/kyma-gke-compass-integration.sh"
-          env:
-            - name: CLOUDSDK_COMPUTE_ZONE
-              value: "europe-west4-b"
-            - name: INPUT_CLUSTER_NAME
-              value: "compass-integration-periodic-on-demand"
-          resources:
-            requests:
-              memory: 200Mi
-              cpu: 80m
-    branches:
-      - ^master$
-    labels:
-      prow.kyma-project.io/slack.skipReport: "true" # this job will be ignored by Slack reporter
-      preset-build-master: "true"
-      preset-kyma-keyring: "true"
-      preset-kyma-encryption-key: "true"
-      preset-kms-gc-project-env: "true"
-      <<: *gke_job_labels_template
-    extra_refs:
-      - <<: *test_infra_ref
-        base_ref: master
-      - org: kfurgol # change to kyma-incubator after testing
-        repo: compass
-        path_alias: github.com/kyma-incubator/compass
-        base_ref: add-cr-with-full-kyma-and-compass # change to master after testing
 
 postsubmits:
   kyma-project/kyma:
@@ -1119,42 +1072,3 @@ periodics:
         requests:
           memory: 200Mi
           cpu: 80m
-
-- name: kyma-gke-compass-integration-periodic
-  cron: "0 0 * * 1-5"
-  decorate: true
-  max_concurrency: 10
-  spec:
-    containers:
-      - image: eu.gcr.io/kyma-project/test-infra/kyma-cluster-infra:v20190528-8897828
-        securityContext:
-          privileged: true
-        command:
-          - "bash"
-        args:
-          - "-c"
-          - "${KYMA_PROJECT_DIR}/test-infra/prow/scripts/cluster-integration/kyma-gke-compass-integration.sh"
-        env:
-          - name: CLOUDSDK_COMPUTE_ZONE
-            value: "europe-west4-b"
-          - name: INPUT_CLUSTER_NAME
-            value: "compass-integration-periodic"
-        resources:
-          requests:
-            memory: 200Mi
-            cpu: 80m
-  branches:
-    - ^master$
-  labels:
-    preset-build-master: "true"
-    preset-kyma-keyring: "true"
-    preset-kyma-encryption-key: "true"
-    preset-kms-gc-project-env: "true"
-    <<: *gke_job_labels_template
-  extra_refs:
-    - <<: *kyma_ref
-      base_ref: master
-    - <<: *test_infra_ref
-      base_ref: master
-    - <<: *compass_ref
-      base_ref: master
